update: added unit tests for stats modal
